+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 - `@InjectWith` and `@UnwrappedInjectWith` parameters property wrappers. [#6](https://github.com/alexruperez/SecurePropertyStorage/pull/6)
 - Instance inject scope.
 - Mock qualifier for testing.
+- `KeychainStorage` access group and synchronizable options. Thanks for the feedback [@JesusAntonioGil](https://github.com/JesusAntonioGil)!
+
+#### Fixed
+- Strong typing on non-`@Store` property wrappers to avoid mixing. Thanks for the feedback [@JesusAntonioGil](https://github.com/JesusAntonioGil)!
 
 [Commits](https://github.com/alexruperez/SecurePropertyStorage/compare/0.3.0...0.3.1)
 
@@ -23,7 +27,7 @@
 ## 0.2.0
 
 #### Added
-- Added `Codable` and `Unwrapped` property wrapper variations. [#3](https://github.com/alexruperez/SecurePropertyStorage/pull/3)
+- Added `Codable` and `Unwrapped` property wrapper variations. [#3](https://github.com/alexruperez/SecurePropertyStorage/pull/3) Thanks for the feedback [@th-mustache-dk](https://github.com/th-mustache-dk)!
 
 [Commits](https://github.com/alexruperez/SecurePropertyStorage/compare/0.1.3...0.2.0)